Some files are open64'd instead of open'd, and some stats are stat'd instead of stat64'd. Fixed regex.

+++ test/test.js
@@ -84,19 +84,23 @@ describe('JailKeeper', function () {
     });
   });
   it('should prevent writing to parent directory', function (done) {
-    fs.unlinkSync('test/parentDirWrite.txt');
-    this.timeout(0);
-    var jail = new JailKeeper();
-    var childProcess = jail.spawn('./writeToParentDirectory', [], { cwd: './test/fixtures' });
-    var jailbroken = false;
-    jail.on('jailbreak', function (code) {
-      jailbroken = true;
-    });
-    jail.on('exit', function (code) {
-      expect(jailbroken).to.be(true);
-      fs.stat('test/parentDirWrite.txt', function (deetz) {
-        expect(deetz).to.be(null);
-        done();
+    fs.stat('test/parentDirWrite.txt', function (deetz) {
+      if (deetz) {
+        fs.unlinkSync('test/parentDirWrite.txt');
+      }
+      this.timeout(0);
+      var jail = new JailKeeper();
+      var childProcess = jail.spawn('./writeToParentDirectory', [], { cwd: './test/fixtures' });
+      var jailbroken = false;
+      jail.on('jailbreak', function (code) {
+        jailbroken = true;
+      });
+      jail.on('exit', function (code) {
+        expect(jailbroken).to.be(true);
+        fs.stat('test/parentDirWrite.txt', function (deetz) {
+          expect(deetz).to.be(null);
+          done();
+        });
       });
     });
   });
